+++ src/components/Lesson/LessonDownloads/LessonDownloads.page.tsx
@@ -382,8 +382,11 @@ export function LessonDownloads(props: LessonDownloadsProps) {
                     />
                   </Box>
                 )}
-
-                <CopyrightNotice showPostAlbCopyright={showPostAlbCopyright} />
+                <Box $mb={56} $mt={16}>
+                  <CopyrightNotice
+                    showPostAlbCopyright={showPostAlbCopyright}
+                  />
+                </Box>
               </>
             )}